+++ src/pages/home/home.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from "react";
-import About from "../../components/details/about/about";
-import Technologies from "../technologies/technologies";
-import Portfolio from "../portfolio/portfolio";
-import Contact from "../contact/contact";
+import About from "../../components/sections/about/about";
+import Technologies from "../../components/sections/technologies/technologies";
+import Portfolio from "../../components/sections/portfolio/portfolio";
+import Contact from "../../components/sections/contact/contact";
 import "./home.css";
 
 const Home = () => {
